+++ avl/_core/__init__.py
@@ -42,7 +42,7 @@ PhaseManager.add_phase("RUN", None, top_down=False)
 PhaseManager.add_phase("REPORT", PhaseManager.get_phase("RUN"), top_down=False)
 
 # Add version
-__version__: str = "0.2.2"
+__version__: str = "0.3.0"
 
 __all__ = [
     "__version__",